code re-formatting

+++ include/number_theory.h
@@ -45,7 +45,7 @@ std::vector<int> x2contfrac(double x, std::size_t n, std::size_t cut = 1e5)
 
     for (std::size_t i = 0; i < n; ++i)
     {
-        result.push_back(std::floor(x));
+        result.push_back(static_cast<int>(std::floor(x)));
         x = 1. / (x - std::floor(x));
         if (!std::isfinite(x) || x > cut)
             return result;
